Admin Action and UI Admin - frontend

+++ backend/src/main/java/com/ou/autorepairshop/dto/QuotationResponse.java
@@ -1,12 +1,13 @@
 package com.ou.autorepairshop.dto;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.List;
 
 public record QuotationResponse(
         Long id,
         String status,
-        double totalPrice,
+        BigDecimal totalPrice,
         LocalDateTime createdAt,
         Long repairOrderId,
         List<QuotationDetailResponse> details
